starting with the big refactor
