+++ components/admin/layers/form/steps/Step1.js
@@ -12,7 +12,7 @@ import Textarea from 'components/form/TextArea';
 import Checkbox from 'components/form/Checkbox';
 import Code from 'components/form/Code';
 import Map from 'components/widgets/editor/map/Map';
-import Legend from 'components/widgets/editor/ui/Legend';
+import Legend from 'components/ui/Legend';
 
 // Utils
 import LayerManager from 'components/widgets/editor/helpers/LayerManager';
