+++ components/app/dashboards/DashboardCard.js
@@ -8,7 +8,7 @@ import Title from 'components/widgets/editor/ui/Title';
 import TextChart from 'components/widgets/charts/TextChart';
 import VegaChart from 'components/widgets/charts/VegaChart';
 import Map from 'components/widgets/editor/map/Map';
-import Legend from 'components/widgets/editor/ui/Legend';
+import Legend from 'components/ui/Legend';
 import Spinner from 'components/widgets/editor/ui/Spinner';
 import Icon from 'components/widgets/editor/ui/Icon';
