+++ components/dashboards/wysiwyg/DashboardWidget.js
@@ -6,7 +6,7 @@ import 'isomorphic-fetch';
 import TextChart from 'components/widgets/charts/TextChart';
 import VegaChart from 'components/widgets/charts/VegaChart';
 import Map from 'components/widgets/editor/map/Map';
-import Legend from 'components/widgets/editor/ui/Legend';
+import Legend from 'components/ui/Legend';
 
 import Title from 'components/widgets/editor/ui/Title';
 import Spinner from 'components/widgets/editor/ui/Spinner';
