+++ components/dashboards/wysiwyg/widget-block/widget-block-component.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import VegaChart from 'components/widgets/charts/VegaChart';
 import TextChart from 'components/widgets/charts/TextChart';
 import Map from 'components/widgets/editor/map/Map';
-import Legend from 'components/widgets/editor/ui/Legend';
+import Legend from 'components/ui/Legend';
 
 import Icon from 'components/ui/Icon';
 import Title from 'components/ui/Title';
