+++ components/ui/LegendType.js
